Add files via upload
All these files are clones of the Extended Oberon System (EOS) by Andreas Pirklbauer, dated February/03.2020. The EOS was previously known as an Experimental Oberon System. It was derived from the 2013 release of Niklaus Wirth's FPGA Oberon System. Please see Andreas' repository for the web pointers and excellent documentation: https://github.com/andreaspirklbauer/Oberon-extended

I downloaded the files on Feb/03. Then I created the SysDef.Mod gathering the memory and register addresses, which were previously scattered all over the sources in the form of hardwired numbers like "12", "24", etc. Andreas helped me understand what these numbers represented. Then I substituted the hardwired literals with symbolic names gathered in one place SysDef.Mod

The goal of this effort is to aid portability to new FPGA boards. It is mandatory to interface the Oberon software to hardware memory locations, but it is not guaranteed that the locations will stay the same as it was the case with the EOS. Having all the relevant memory addresses in one file will allow to change just a single file, recompile the software, and run it on a new FPGA board.

Status: I have not compiled this software yet because I am learning how to use RISC5 emulators. Also, the files are now in the BlackBox format to help with source formatting using colors and fonts. These files need be saved as ASCII or Oberon Text before being imported into the RISC5 emulator. The ASCII versions are provided. I will make these into the Oberon Text format when I actually begin to work under the emulated Oberon System.