Add files via upload
These are the same files as in the BlackBox format repository, but in pure ASCII. No colors, no fonts, no formatting. Additionally, there are the fonts. This is a complete set of files comprising the Extended Oberon System as of Feb/03/2020.

Note that the literal memory addresses like "12" or "24" were replaced with symbolic names defined in SysDef.Mod. Additionally, each file starts with an exported VERSION string literal.